+++ base/deprecated.jl
@@ -1691,7 +1691,7 @@ export hex2num
 # deprecations for filter[!] with 2-arg functions are in associative.jl
 
 # PR 23341
-@deprecate diagm(A::SparseMatrixCSC) diagm(sparsevec(A))
+@deprecate diagm(A::SparseMatrixCSC) spdiagm(sparsevec(A))
 
 # END 0.7 deprecations